requirements.txt
python-dotenv>=1.0.0
openai-agents>=0.0.1
gradio>=6.10.0
